+++ python-package/tests/test_10_nichtparasoup/test_imagecrawlers/test_instagram.py
@@ -56,8 +56,8 @@ class _InstagramFileFetcher(FileFetcher):
 
 _FILE_FETCHER = _InstagramFileFetcher({  # relative to './testdata_instagram'
     '/': 'index.html',
-    '/static/bundles/metro/ProfilePageContainer.js/e243abb1f92c.js': 'e243abb1f92c.js',
     '/static/bundles/metro/TagPageContainer.js/1bad9348735e.js': '1bad9348735e.js',
+    '/static/bundles/metro/Consumer.js/ebbdfced63f8.js': 'ebbdfced63f8.js',
     '/graphql/query/?query_hash=f0986789a5c5d17c2400faebf16efd0d&'
     'variables=%7B%22first%22%3A+1%2C+%22after%22%3A+%22%22%2C+%22tag_name%22%3A+%22foo%22%7D':
         'query_hash=f0986789a5c5d17c2400faebf16efd0d&variables={first-1,after,tag_name-foo}',
@@ -70,32 +70,31 @@ _FILE_FETCHER = _InstagramFileFetcher({  # relative to './testdata_instagram'
     '/graphql/query/?query_hash=174a5243287c5f3a7de741089750ab3b&'
     'variables=%7B%22first%22%3A+5%2C+%22after%22%3A+%22%22%2C+%22tag_name%22%3A+%22foo%22%7D':
         'query_hash=174a5243287c5f3a7de741089750ab3b&variables={first-5,after,tag_name-foo}',
+    '/natgeo/': 'natgeo',
     '/natgeo/?__a=1': 'natgeo.__a=1',
-    '/graphql/query/?query_hash=2c5d4d8b70cad329c4a6ebe3abb6eedd&'
+    '/graphql/query/?query_hash=51fdd02b67508306ad4484ff574a0b62&'
     'variables=%7B%22first%22%3A+1%2C+%22after%22%3A+%22%22%2C+%22id%22%3A+%22787132%22%7D':
-        'query_hash=2c5d4d8b70cad329c4a6ebe3abb6eedd&variables={first-1,after,id-787132}',
+        'query_hash=51fdd02b67508306ad4484ff574a0b62&variables={first-1,after,id-787132}',
     '/graphql/query/?query_hash=ff260833edf142911047af6024eb634a&'
     'variables=%7B%22first%22%3A+1%2C+%22after%22%3A+%22%22%2C+%22id%22%3A+%22787132%22%7D':
         'query_hash=ff260833edf142911047af6024eb634a&variables={first-1,after,id-787132}',
     '/graphql/query/?query_hash=f0986789a5c5d17c2400faebf16efd0d&'
     'variables=%7B%22first%22%3A+1%2C+%22after%22%3A+%22%22%2C+%22id%22%3A+%22787132%22%7D':
         'query_hash=f0986789a5c5d17c2400faebf16efd0d&variables={first-1,after,id-787132}',
-    '/graphql/query/?query_hash=8c86fed24fa03a8a2eea2a70a80c7b6b&'
+    '/graphql/query/?query_hash=97b41c52301f77ce508f55e66d17620e&'
     'variables=%7B%22first%22%3A+1%2C+%22after%22%3A+%22%22%2C+%22id%22%3A+%22787132%22%7D':
-        'query_hash=8c86fed24fa03a8a2eea2a70a80c7b6b&variables={first-1,after,id-787132}',
-    '/graphql/query/?query_hash=2c5d4d8b70cad329c4a6ebe3abb6eedd&'
+        'query_hash=97b41c52301f77ce508f55e66d17620e&variables={first-1,after,id-787132}',
+    '/graphql/query/?query_hash=51fdd02b67508306ad4484ff574a0b62&'
     'variables=%7B%22first%22%3A+5%2C+%22after%22%3A+%22%22%2C+%22id%22%3A+%22787132%22%7D':
-        'query_hash=2c5d4d8b70cad329c4a6ebe3abb6eedd&variables={first-5,after,id-787132}',
+        'query_hash=51fdd02b67508306ad4484ff574a0b62&variables={first-5,after,id-787132}',
 }, base_url='https://www.instagram.com/', base_dir=_InstagramFileFetcher.TESTDATA_PATH)
 
 _QUERYHASHES_EXPECTED_TAG = {'f0986789a5c5d17c2400faebf16efd0d',
                              'ff260833edf142911047af6024eb634a',
                              '174a5243287c5f3a7de741089750ab3b'}
 
-_QUERYHASHES_EXPECTED_PROFILE = {'f0986789a5c5d17c2400faebf16efd0d',
-                                 'ff260833edf142911047af6024eb634a',
-                                 '2c5d4d8b70cad329c4a6ebe3abb6eedd',
-                                 '8c86fed24fa03a8a2eea2a70a80c7b6b'}
+_QUERYHASHES_EXPECTED_PROFILE = {'97b41c52301f77ce508f55e66d17620e',
+                                 '51fdd02b67508306ad4484ff574a0b62'}
 
 
 class InstagramQueryHashFinderTest(unittest.TestCase):
@@ -368,9 +367,45 @@ def test_instagram_hashtag_loader() -> None:
     ImageCrawlerLoaderTest().check('InstagramHashtag', InstagramHashtag)
 
 
+class TestInstagramProfileConfig:
+
+    def test_neither_name_not_id(self) -> None:
+        with pytest.raises(KeyError, match=r'either .* required'):
+            InstagramProfile.check_config(dict())
+
+    def test_name_and_id(self) -> None:
+        with pytest.raises(KeyError, match=r'either .* required'):
+            InstagramProfile.check_config(dict(user_name='foo', profile_id=123))
+
+    def test_name_wrong_type(self) -> None:
+        for wrong in [False, 23, 4.2, [], (), {}, self]:
+            with pytest.raises(TypeError, match=r'user_name .* is not str'):
+                InstagramProfile._check_config_name(wrong)
+
+    def test_name_wrong_value(self) -> None:
+        with pytest.raises(ValueError, match=r'user_name .* is empty'):
+            InstagramProfile._check_config_name('')
+
+    def test_name_correct(self) -> None:
+        InstagramProfile._check_config_name('foo')
+
+    def test_id_wrong_type(self) -> None:
+        for wrong in [False, '23', 4.2, [], (), {}, self]:
+            with pytest.raises(TypeError, match=r'profile_id .* is not int'):
+                InstagramProfile._check_config_id(wrong)
+
+    def test_id_wrong_value(self) -> None:
+        for wrong in [-1, 0]:
+            with pytest.raises(ValueError, match=r'profile_id .* is <= 0'):
+                InstagramProfile._check_config_id(wrong)
+
+    def test_id_correct(self) -> None:
+        InstagramProfile._check_config_id(23)
+
+
 class InstagramProfileTest(unittest.TestCase):
     _PROFILE_ID = '787132'
-    _QUERY_HASH = '2c5d4d8b70cad329c4a6ebe3abb6eedd'
+    _QUERY_HASH = '51fdd02b67508306ad4484ff574a0b62'
 
     def setUp(self) -> None:
         InstagramProfile._query_hash = None